soft server socket max limit, test refactoring
