+++ lib/queue.js
@@ -101,3 +101,5 @@ class PeerQueue extends EventEmitter {
 function toID (peer) {
   return peer.host + ':' + peer.port
 }
+
+module.exports.PeerQueue = PeerQueue
