+++ package.json
@@ -16,8 +16,8 @@
     "tap": "^14.1.11"
   },
   "scripts": {
-    "test": "tap -R classic test/*.test.js && standard --fix",
-    "cov": "tap -R classic --coverage-report=html test/*.test.js"
+    "test": "tap -j 1 -R classic test/*.test.js && standard --fix",
+    "cov": "tap -j 1 -R classic --coverage-report=html test/*.test.js"
   },
   "repository": {
     "type": "git",
